Switch to the C API of SPIRV-Cross
Also moves the remap (e.g. stripping of symbols) support to
its own file. (useful to prevent header clashes between glslang
and SPIRV-Cross)

The C API of SPIRV-Cross is expected to be more stable than
the C++ one (the latter comes with no guarantees). It is worth
noting that full behavioral compatibility when it comes to the
generated shader code is not part of the compatibility promise,
but being able to build on a stable C API is already a good step
forward.

Task-number: QTBUG-78591
Change-Id: I9a91bb60a2f29bf91670f00659824636f6b6fc2a
Reviewed-by: Christian Strømme <christian.stromme@qt.io>