Drop binding qualifiers from samplers in GLSL
Enabling GL_ARB_shading_language_420pack (for core-compatible shaders)
is not quite ok when the OpenGL implementation (e.g. the one on macOS)
does not support this extension.

The QRhi OpenGL backend does not rely on bindings for samplers in the
first place (and uniform blocks are converted to plain uniforms). So
there is no purpose for the binding points in the GLSL code.

This fixes operating on macOS with the 4.1 core profile context.
(2.1 worked in any case, as this flag and the concept of layout
qualifiers are not relevant for <= 120 shaders)

Change-Id: Iaec2130aac5fef7c3d8656718bd9ccf418299580
Fixes: QTBUG-80690
Reviewed-by: Johan Helsing <johan.helsing@qt.io>